--- .eslintignore
@@ -1,1 +1,0 @@
-src/__tests__/*.js
