+++ __tests__/defaults/retry.js
@@ -1,4 +1,4 @@
-import retry from '../../defaults/retry';
+import retry from '../../src/defaults/retry';
 
 test('no arguments', () => {
   expect(retry()).toBeNull();
